Change 0.1
Added piece set. This contains all the different rotations of each piece. Since I'm not trying to make a full game of tetris, simply make it work with machine learning, doing something like this makes it easier to handle than having the actual rotation system implemented. This does, for the time being, remove the possibility of doing tucks and spins, but I'm okay with that since my goal for now is very simple. Doing rotations would make this project about 10 times more difficult and I'm not too interested in doing that, at least not yet.