Change 0.3
Added piece randomization and the rest of the framework for the main class. I do need to add resets and adjustments to the neural network before that is totally done, but for the most part it is finished. Next and final step is just setting up the neural network itself and its inputs. This shouldn't be a huge undertaking, although calculating the different inputs that I want will be a little tricky.