Version 2
This might be it for this project, which is weird to say the same day I started it. It is literally perfect. I left it running while I ate dinner (with a decreased tick speed) and I've just come back to it having cleared 40,000 lines. I can't think of a feasible way of improving it, at least with this structure. For the time being, it works.